SCRIPTS: more bash-to-POSIX stuff

+++ .config/i3/volume-watcher.sh
@@ -1,11 +1,13 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 
 set -euo pipefail
 
 # This script uses pw-mon to watch for changes in the volume and mute status
 # and signals i3blocks to update the volume block
 
-pkill -f "pw-mon"
+if pgrep -f "pw-mon"; then
+    pkill -f "pw-mon"
+fi
 
 pw-mon | grep --line-buffered "node.name" | while read -r _; do
     pkill -SIGRTMIN+10 i3blocks # signal i3blocks to update the volume block
